Use the xenial version of osm2pgsql
After the appearance of yakkety yak, backportpackage began trying to
build 16.10's osm2pgsql (0.90), which includes dependencies that can't
be resolved. This pins the backported version to xenial, which is 0.88.

The failure to backport osm2pgsql-0.90 was resulting in the version
shipped in trusty (0.82) being installed, which didn't include support
(at least in the way we expected) to --extra-attributes.

Fixes AmericanRedCross/posm#129